Restructure the project: More crates for better code reuse
Our old project layout was fine so far, but had it's weaknesses. With this new
layout we have the following elements:
- server/src/lib.rs: (private) library that contains all the old server code
- server/server.rs: server executable that uses said library
- server/examples/*.rs: Different executables testing server code
- src/lib.rs: API library for public use (helps communicating with our server)
- src/client/main.rs: CLI client for our server

I moved many files around but didn't really change anything. That means that
the code is not really prepared for code reuse or anything. Many things need
to move around, before the actual new structure is ready.

Some things needed to be tweaked in order to function properly: The logger won't
cut of parts of the file path (not every path begins with src/ anymore), the
.gitignore got some new entries and check-style will check all .rs files
now (I hope that won't break travis...)
