+++ .gitignore
@@ -1,9 +1,12 @@
 # Generated by Cargo
 /target/
+/server/target/
 /Cargo.lock
+/server/Cargo.lock
 
 # Generated by sublime
 *.sublime-workspace
 
 # Generated by our server
 /log.txt
+/server/log.txt
