+++ server/Cargo.toml
@@ -1,15 +1,11 @@
 [package]
-name = "uosql"
+name = "server"
 version = "0.1.0"
 authors = ["Lukas Kalbertodt <lukas.kalbertodt@gmail.com>"]
 
-[lib]
-name = "uosql"
-path = "src/lib.rs"
-
 [[bin]]
-name = "uosql-client"
-path = "src/client/main.rs"
+name = "uosql-server"
+path = "server.rs"
 
 [dependencies]
 log = "*"
@@ -18,6 +14,3 @@ byteorder = "0.3"
 rustc-serialize = "0.3"
 bincode = "0.4.0"
 docopt = "0.6"
-
-[dependencies.server]
-path = "server"
