+++ server/config.json
