+++ server/examples/parser.rs
@@ -1,8 +1,8 @@
 ///  Program for testing and playing with the parser
 ///
 
-extern crate uosql;
-use uosql::parse;
+extern crate server;
+use server::parse;
