+++ server/src/auth.rs
