+++ src/lib.rs
@@ -1,0 +1,7 @@
+#[macro_use]
+extern crate server;
+
+pub use server::*;
+
+#[test]
+fn it_works() {}
